CRUD em memoria volatil terminado. Vamos comecar a usar o Firebase

+++ src/pages/detail/detail.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, AlertController } from 'ionic-angular';
 
 import {NoteService} from '../../app/note.service'
 /**
@@ -17,20 +17,52 @@ import {NoteService} from '../../app/note.service'
 export class DetailPage {
 
   note;
+  newNoteFlag = false;
   constructor(public navCtrl: NavController,
            public navParams: NavParams,
-            private noteService: NoteService) {
+            private noteService: NoteService,
+            private alertCtrl : AlertController) {
     this.note = this.navParams.get("noteParam");
+    if (!this.note){
+      this.note = {
+        title: "",
+        date: "",
+        content: ""
+      };
+      this.newNoteFlag = true;
+    }
     console.log ('nav-param', this.note);
   }
 
   onTrash (){
-    this.noteService.removeNote(this.note);
-    this.navCtrl.pop();
+    let confirm = this.alertCtrl.create({
+      title: "Delete?",
+      message: `Are you sure you want to delete this note? ${this.note.title}`,
+      buttons: [
+        {
+          text: "Cancel"
+        },
+        {
+          text: "Confirm",
+          handler: () =>{
+            this.noteService.removeNote(this.note);
+            this.navCtrl.pop();
+          }
+        }
+      ]
+    }); 
+    confirm.present();   
   }
 
   ionViewDidLoad() {
     console.log('ionViewDidLoad DetailPage');
+  }
+
+  ionViewWillLeave(){
+    if (this.newNoteFlag)
+      if (this.note.date != ""
+        && this.note.title != "" && this.note.content != "")
+          this.noteService.addNote(this.note);
   }
 
 }
